refactor: move injection to di package
